feat: refactor createProduct lambda to request dynamodb
