+++ product-service/package.json
@@ -30,6 +30,7 @@
     "@aws-sdk/lib-dynamodb": "^3.598.0",
     "aws-cdk-lib": "^2.145.0",
     "constructs": "^10.3.0",
+    "joi": "^17.13.1",
     "source-map-support": "^0.5.21"
   }
 }
